+++ geotrek/api/v2/views/outdoor.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from django.db.models import F
+from django.utils.translation import activate
 
 from geotrek.api.v2 import serializers as api_serializers, \
     filters as api_filters, viewsets as api_viewsets
@@ -10,9 +11,12 @@ from geotrek.outdoor import models as outdoor_models
 class SiteViewSet(api_viewsets.GeotrekGeometricViewset):
     filter_backends = api_viewsets.GeotrekGeometricViewset.filter_backends + (api_filters.GeotrekSiteFilter,)
     serializer_class = api_serializers.SiteSerializer
-    queryset = outdoor_models.Site.objects \
-        .annotate(geom_transformed=Transform(F('geom'), settings.API_SRID)) \
-        .order_by('pk')  # Required for reliable pagination
+
+    def get_queryset(self):
+        activate(self.request.GET.get('language'))
+        return outdoor_models.Site.objects \
+            .annotate(geom_transformed=Transform(F('geom'), settings.API_SRID)) \
+            .order_by('name')  # Required for reliable pagination
 
 
 class OutdoorPracticeViewSet(api_viewsets.GeotrekGeometricViewset):
@@ -44,6 +48,9 @@ class RatingViewSet(api_viewsets.GeotrekViewSet):
 class CourseViewSet(api_viewsets.GeotrekGeometricViewset):
     filter_backends = api_viewsets.GeotrekGeometricViewset.filter_backends + (api_filters.GeotrekCourseFilter,)
     serializer_class = api_serializers.CourseSerializer
-    queryset = outdoor_models.Course.objects \
-        .annotate(geom_transformed=Transform(F('geom'), settings.API_SRID)) \
-        .order_by('pk')  # Required for reliable pagination
+
+    def get_queryset(self):
+        activate(self.request.GET.get('language'))
+        return outdoor_models.Course.objects \
+            .annotate(geom_transformed=Transform(F('geom'), settings.API_SRID)) \
+            .order_by('name')  # Required for reliable pagination